[ADM-240] Extract methods in GenerateSprintReporterService.ts to Jira.ts (#31)
* refactor: extract all methods in GenerateSprintReporterService.ts to Jira.ts

* refactor: extract all the tests in GenerateSprintReporterService.test.ts to jira.test.ts
