+++ backend/tests/services/kanban/Jira/Jira.test.ts
@@ -13,39 +13,170 @@ import NonDoneJiraCardCycleTime from "../../../fixture/NonDoneJiraCardCycleTime.
 import StatusData from "../../../fixture/StatusData.json";
 import { StoryPointsAndCycleTimeRequest } from "../../../../src/contract/kanban/KanbanStoryPointParameterVerify";
 import { Sprint } from "../../../../src/models/kanban/Sprint";
-import { CycleTimeInfo } from "../../../../src/contract/kanban/KanbanStoryPointResponse";
+import {
+  CycleTimeInfo,
+  JiraCardResponse,
+} from "../../../../src/contract/kanban/KanbanStoryPointResponse";
 import { StatusSelf } from "../../../../src/models/kanban/JiraBoard/StatusSelf";
 import { StatusCategory } from "../../../../src/models/kanban/JiraBoard/StatusSelf";
-
 import axios from "axios";
 import {
   HistoryDetail,
   JiraCardHistory,
   Status,
 } from "../../../../src/models/kanban/JiraCardHistory";
+import {
+  JiraCard,
+  JiraCardField,
+} from "../../../../src/models/kanban/JiraCard";
+import { JiraBlockReasonEnum } from "../../../../src/models/kanban/JiraBlockReasonEnum";
+import {
+  CardCycleTime,
+  StepsDay,
+} from "../../../../src/models/kanban/CardCycleTime";
+import { SprintStatistics } from "../../../../src/models/kanban/SprintStatistics";
 
 const jira = new Jira("testToken", "domain");
 const jiraTest = Object.getPrototypeOf(jira);
-describe("get sprints data by domain name and boardId", () => {
-  const storyPointsAndCycleTimeRequest = new StoryPointsAndCycleTimeRequest(
-    "testToken",
-    "jira",
-    "domain",
-    "project",
-    "2",
-    ["Done"],
-    1589080044000,
-    1589944044000,
-    [
-      {
-        key: "customfield_10016",
-        name: "Story point estimate",
-        flag: true,
-      },
-    ],
-    false
-  );
 
+const sprint1Name = "test Sprint 1";
+const sprint2Name = "test Sprint 2";
+const sprint3Name = "test Sprint 3";
+
+const jiraCardField1: JiraCardField = new JiraCardField();
+const jiraCardField2: JiraCardField = new JiraCardField();
+const jiraCardField3: JiraCardField = new JiraCardField();
+
+const activeSprint = new Sprint(
+  1,
+  "active",
+  sprint1Name,
+  "2020-07-22T01:48:39.455Z",
+  "2020-08-09T03:20:37.000Z"
+);
+const closedSprint = new Sprint(
+  4,
+  "closed",
+  sprint2Name,
+  "2020-05-26T03:20:43.632Z",
+  "2020-08-05T01:48:37.000Z",
+  "2020-08-22T01:49:26.508Z"
+);
+
+const futureSprint = new Sprint(9, "future", sprint3Name);
+
+const sprints: Sprint[] = [activeSprint, closedSprint, futureSprint];
+const activeAndClosedSprints = [activeSprint, closedSprint];
+
+jiraCardField1.sprint = sprint1Name;
+jiraCardField1.label = JiraBlockReasonEnum.DEPENDENCIES_NOT_WORK;
+
+jiraCardField2.sprint = sprint2Name;
+jiraCardField2.label = JiraBlockReasonEnum.TAKE_LEAVE;
+
+const jiraCard1: JiraCard = { fields: jiraCardField1, key: "" };
+const jiraCard2: JiraCard = { fields: jiraCardField2, key: "" };
+const jiraCard3: JiraCard = { fields: jiraCardField3, key: "" };
+
+const cycleTimeArray1: CycleTimeInfo[] = [
+  { column: "DOING", day: 1 },
+  { column: "WAIT", day: 2 },
+  { column: "TEST", day: 3 },
+  { column: "BLOCKED", day: 4 },
+  { column: "REVIEW", day: 5 },
+];
+const cycleTimeArray2: CycleTimeInfo[] = [
+  { column: "DOING", day: 2 },
+  { column: "WAIT", day: 3 },
+  { column: "TEST", day: 4 },
+  { column: "BLOCKED", day: 5 },
+  { column: "REVIEW", day: 6 },
+];
+const cycleTimeArray3: CycleTimeInfo[] = [
+  { column: "DOING", day: 1 },
+  { column: "WAIT", day: 8 },
+  { column: "TEST", day: 4 },
+  { column: "BLOCKED", day: 9 },
+  { column: "REVIEW", day: 6 },
+];
+
+const stepsDay: StepsDay = {
+  analyse: 0,
+  development: 0,
+  waiting: 0,
+  testing: 0,
+  blocked: 1,
+  review: 0,
+};
+
+const cardCycleTime1: CardCycleTime = {
+  name: "Card 1",
+  total: 2,
+  steps: stepsDay,
+};
+const cardCycleTime2: CardCycleTime = {
+  name: "Card 2",
+  total: 5,
+  steps: stepsDay,
+};
+const cardCycleTime3: CardCycleTime = {
+  name: "Card 3",
+  total: 6.6,
+  steps: stepsDay,
+};
+
+const card1 = new JiraCardResponse(
+  jiraCard1,
+  cycleTimeArray1,
+  [],
+  cardCycleTime1
+);
+const card2 = new JiraCardResponse(
+  jiraCard2,
+  cycleTimeArray2,
+  [],
+  cardCycleTime2
+);
+const card3 = new JiraCardResponse(
+  jiraCard3,
+  cycleTimeArray3,
+  [],
+  cardCycleTime3
+);
+
+const cardList1 = [card1, card2, card3];
+const cardList2 = [card1, card2];
+const cardList3 = [card1];
+const emptyCardList: JiraCardResponse[] = [];
+
+const mapSprintCards = new Map<string, JiraCardResponse[]>([
+  [sprint1Name, cardList1],
+  [sprint2Name, cardList2],
+  [sprint3Name, cardList3],
+]);
+const mapSprintEmptyCards = new Map<string, JiraCardResponse[]>([
+  [sprint1Name, emptyCardList],
+]);
+const storyPointsAndCycleTimeRequest = new StoryPointsAndCycleTimeRequest(
+  "testToken",
+  "jira",
+  "domain",
+  "project",
+  "2",
+  ["Done"],
+  1589080044000,
+  1589944044000,
+  [
+    {
+      key: "customfield_10016",
+      name: "Story point estimate",
+      flag: true,
+    },
+  ],
+  false
+);
+
+describe("get sprints data by domain name and boardId", () => {
   const sprint1 = new Sprint(
     1,
     "closed",
@@ -178,13 +309,13 @@ describe("get query status", () => {
   afterEach(() => sinon.restore());
   it("should return query status data", async () => {
     mock
-      .onGet(`https://domain.atlassian.net/rest/api/2/status/10006`, {
-        headers: { Authorization: `testToken` },
+      .onGet("https://domain.atlassian.net/rest/api/2/status/10006", {
+        headers: { Authorization: "testToken" },
       })
       .reply(200, StatusData);
     const result: StatusSelf = await jiraTest.constructor.queryStatus(
       "https://domain.atlassian.net/rest/api/2/status/10006",
-      `testToken`
+      "testToken"
     );
     expect(result.statusCategory.key).equal("indeterminate");
   });
@@ -409,7 +540,7 @@ describe("get all done cards", () => {
       .reply(200, JiraCards);
 
     jiraTest.httpClient = axios.create({
-      baseURL: `https://domain.atlassian.net/rest/agile/1.0/board`,
+      baseURL: "https://domain.atlassian.net/rest/agile/1.0/board",
     });
     jiraTest.httpClient.defaults.headers.common["Authorization"] = "testToken";
     jiraTest.queryCount = 100;
@@ -457,7 +588,7 @@ describe("get all non done cards for active sprint", () => {
       .reply(200, NonDoneJiraCard);
 
     jiraTest.httpClient = axios.create({
-      baseURL: `https://domain.atlassian.net/rest/agile/1.0/board`,
+      baseURL: "https://domain.atlassian.net/rest/agile/1.0/board",
     });
     jiraTest.httpClient.defaults.headers.common["Authorization"] = "testToken";
     jiraTest.queryCount = 100;
@@ -504,7 +635,7 @@ describe("get all non done cards for kanban", () => {
       .reply(200, JiraCards);
 
     jiraTest.httpClient = axios.create({
-      baseURL: `https://domain.atlassian.net/rest/agile/1.0/board`,
+      baseURL: "https://domain.atlassian.net/rest/agile/1.0/board",
     });
     jiraTest.httpClient.defaults.headers.common["Authorization"] = "testToken";
     jiraTest.queryCount = 100;
@@ -760,7 +891,7 @@ describe("make page query", () => {
     const cards: any = [];
     jiraTest.queryCount = 100;
     jiraTest.httpClient = axios.create({
-      baseURL: `https://domain.atlassian.net/rest/agile/1.0/board`,
+      baseURL: "https://domain.atlassian.net/rest/agile/1.0/board",
     });
     jiraTest.httpClient.defaults.headers.common["Authorization"] = "testToken";
     mock
@@ -774,5 +905,382 @@ describe("make page query", () => {
       .reply(200, JiraCards);
     await jiraTest.pageQuerying(total, jql, cards, 2);
     expect(cards.length).equal(2);
+  });
+});
+
+describe("map cards by sprint name", () => {
+  const jiraCardField1: JiraCardField = new JiraCardField();
+  const jiraCardField2: JiraCardField = new JiraCardField();
+  const jiraCard1: JiraCard = { fields: jiraCardField1, key: "" };
+  const jiraCard2: JiraCard = { fields: jiraCardField2, key: "" };
+  const card1 = new JiraCardResponse(jiraCard1, cycleTimeArray1);
+  const card2 = new JiraCardResponse(jiraCard2, cycleTimeArray2);
+  const cardList = [card1, card2];
+
+  it("should return a map when all cards have sprint names", () => {
+    jiraCardField1.sprint = sprint1Name;
+    jiraCardField2.sprint = sprint2Name;
+
+    const mapSprintCards = jiraTest.mapCardsBySprintName(cardList);
+
+    expect(mapSprintCards).deep.equal(
+      new Map<string, JiraCardResponse[]>([
+        [sprint1Name, [card1]],
+        [sprint2Name, [card2]],
+      ])
+    );
+  });
+
+  it("should return a map with one entry when a part of the cards have sprint names", () => {
+    jiraCardField1.sprint = sprint1Name;
+    jiraCardField2.sprint = "";
+
+    const mapSprintCards = jiraTest.mapCardsBySprintName(cardList);
+
+    expect(mapSprintCards).deep.equal(
+      new Map<string, JiraCardResponse[]>([[sprint1Name, [card1]]])
+    );
+  });
+
+  it("should return empty map when all cards do not have any sprint name", () => {
+    jiraCardField1.sprint = "";
+    jiraCardField2.sprint = "";
+
+    const mapSprintCards = jiraTest.mapCardsBySprintName(cardList);
+
+    expect(mapSprintCards).deep.equal(new Map<string, JiraCardResponse[]>());
+  });
+
+  it("should return empty map when the card list is empty", () => {
+    const mapSprintCards = jiraTest.mapCardsBySprintName(emptyCardList);
+
+    expect(mapSprintCards).deep.equal(new Map<string, JiraCardResponse[]>());
+  });
+});
+describe("calculate percentage of different block reasons in the latest sprint", () => {
+  it("should return correct blocked reason percentage when there are matched cards in sprint", () => {
+    const blockedPercentageByReason = jiraTest.calculateBlockReasonPercentage(
+      activeAndClosedSprints,
+      mapSprintCards
+    );
+
+    expect(
+      blockedPercentageByReason.get(JiraBlockReasonEnum.DEPENDENCIES_NOT_WORK)
+    ).equal(0.07);
+    expect(blockedPercentageByReason.get(JiraBlockReasonEnum.TAKE_LEAVE)).equal(
+      0.07
+    );
+    expect(blockedPercentageByReason.get(JiraBlockReasonEnum.OTHERS)).equal(
+      0.07
+    );
+  });
+
+  it("should return 0 when there is not any matched card in sprint", () => {
+    const mapLatestSprintEmptyCards = new Map<string, JiraCardResponse[]>([
+      [sprint1Name, emptyCardList],
+      [sprint2Name, cardList2],
+      [sprint3Name, cardList3],
+    ]);
+    const blockedPercentageByReason = jiraTest.calculateBlockReasonPercentage(
+      activeAndClosedSprints,
+      mapLatestSprintEmptyCards
+    );
+
+    expect(
+      blockedPercentageByReason.get(JiraBlockReasonEnum.DEPENDENCIES_NOT_WORK)
+    ).equal(0);
+    expect(blockedPercentageByReason.get(JiraBlockReasonEnum.TAKE_LEAVE)).equal(
+      0
+    );
+    expect(blockedPercentageByReason.get(JiraBlockReasonEnum.OTHERS)).equal(0);
+    expect(
+      blockedPercentageByReason.get(JiraBlockReasonEnum.QUESTION_TO_BE_ANSWERED)
+    ).equal(0);
+  });
+});
+
+describe("calculate standard deviation", () => {
+  it("should return correct standard deviation and avgerage when there are matched cards in sprint", () => {
+    const expected: Map<string, any> = new Map<string, any>();
+    expected.set(sprint1Name, { standardDeviation: 1.91, average: 4.53 });
+    expected.set(sprint2Name, { standardDeviation: 1.5, average: 3.5 });
+    expected.set(sprint3Name, { standardDeviation: 0, average: 2 });
+
+    const mapSprintStandardDeviation =
+      jiraTest.calculateStandardDeviation(mapSprintCards);
+
+    expect(mapSprintStandardDeviation).deep.equal(expected);
+  });
+
+  it("should return 0 when there is not any matched card in sprint", () => {
+    const expected: Map<string, any> = new Map<string, any>();
+    expected.set(sprint1Name, { standardDeviation: 0, average: 0 });
+
+    const mapSprintStandardDeviation =
+      jiraTest.calculateStandardDeviation(mapSprintEmptyCards);
+
+    expect(mapSprintStandardDeviation).deep.equal(expected);
+  });
+});
+
+describe("calculate blocked and developing percentage", () => {
+  it("should return correct blocked and developing percentage when there are matched cards in sprint", () => {
+    const mapIterationBlockedPercentage =
+      jiraTest.calculateBlockedAndDevelopingPercentage(mapSprintCards);
+
+    const expected: Map<string, any> = new Map<string, any>();
+    expected.set(sprint1Name, {
+      blockedPercentage: 0.22,
+      developingPercentage: 0.78,
+    });
+    expected.set(sprint2Name, {
+      blockedPercentage: 0.29,
+      developingPercentage: 0.71,
+    });
+    expected.set(sprint3Name, {
+      blockedPercentage: 0.5,
+      developingPercentage: 0.5,
+    });
+
+    expect(mapIterationBlockedPercentage).deep.equal(expected);
+  });
+
+  it("should return 0% blocked percentage and 100% developing percentage when there is not any matched card in sprint", () => {
+    const mapIterationBlockedPercentage =
+      jiraTest.calculateBlockedAndDevelopingPercentage(mapSprintEmptyCards);
+
+    const expected: Map<string, any> = new Map<string, number>();
+    expected.set(sprint1Name, {
+      blockedPercentage: 0,
+      developingPercentage: 1,
+    });
+
+    expect(mapIterationBlockedPercentage).deep.equal(expected);
+  });
+});
+
+describe("calculate the number of completed cards in every sprint", () => {
+  it("should return the correct number of completed cards when there are matched cards in sprint", () => {
+    const mapSprintCards = new Map<string, JiraCardResponse[]>([
+      [sprint1Name, cardList1],
+      [sprint2Name, cardList2],
+    ]);
+
+    const mapSprintCompletedCardsNumber =
+      jiraTest.calculateCompletedCardsNumber(mapSprintCards);
+
+    const expected: Map<string, number> = new Map<string, number>();
+    expected.set(sprint1Name, 3);
+    expected.set(sprint2Name, 2);
+
+    expect(mapSprintCompletedCardsNumber).deep.equal(expected);
+  });
+
+  it("should return 0 when the cardList is empty", () => {
+    const mapSprintCards = new Map<string, JiraCardResponse[]>([
+      [sprint1Name, emptyCardList],
+    ]);
+
+    const mapSprintCompletedCardsNumber =
+      jiraTest.calculateCompletedCardsNumber(mapSprintCards);
+    const expected: Map<string, number> = new Map<string, number>();
+    expected.set(sprint1Name, 0);
+
+    expect(mapSprintCompletedCardsNumber).deep.equal(expected);
+  });
+});
+
+describe("generate Jira sprint statistics", () => {
+  it("should return the Jira sprint statistics when statistic data are not empty", () => {
+    const sprintBlockReasonPercentageMap = new Map<string, number>([
+      [JiraBlockReasonEnum.DEPENDENCIES_NOT_WORK, 0.07],
+      [JiraBlockReasonEnum.TAKE_LEAVE, 0.07],
+      [JiraBlockReasonEnum.OTHERS, 0.07],
+    ]);
+    const sprintBlockPercentageMap = new Map<string, any>([
+      [sprint1Name, { blockedPercentage: 0.22, developingPercentage: 0.78 }],
+      [
+        sprint2Name,
+        {
+          blockedPercentage: 0.29,
+          developingPercentage: 0.71,
+        },
+      ],
+      [
+        sprint3Name,
+        {
+          blockedPercentage: 0.5,
+          developingPercentage: 0.5,
+        },
+      ],
+    ]);
+    const sprintStandardDeviationMap = new Map<string, any>([
+      [sprint1Name, { standardDeviation: 1.91, average: 4.53 }],
+      [sprint2Name, { standardDeviation: 1.5, average: 3.5 }],
+      [sprint3Name, { standardDeviation: 0, average: 2 }],
+    ]);
+    const sprintCompletedCardsNumberMap = new Map<string, number>([
+      [sprint1Name, 3],
+      [sprint2Name, 2],
+    ]);
+
+    const sprintStartistics = jiraTest.generateJiraSprintStatistics(
+      sprintCompletedCardsNumberMap,
+      sprintBlockPercentageMap,
+      sprintStandardDeviationMap,
+      sprintBlockReasonPercentageMap
+    );
+
+    const expected = new SprintStatistics(
+      [
+        { sprintName: sprint1Name, value: 3 },
+        { sprintName: sprint2Name, value: 2 },
+      ],
+      [
+        {
+          sprintName: sprint1Name,
+          value: { standardDeviation: 1.91, average: 4.53 },
+        },
+        {
+          sprintName: sprint2Name,
+          value: { standardDeviation: 1.5, average: 3.5 },
+        },
+        {
+          sprintName: sprint3Name,
+          value: { standardDeviation: 0, average: 2 },
+        },
+      ],
+      [
+        {
+          sprintName: sprint1Name,
+          value: { blockedPercentage: 0.22, developingPercentage: 0.78 },
+        },
+        {
+          sprintName: sprint2Name,
+          value: {
+            blockedPercentage: 0.29,
+            developingPercentage: 0.71,
+          },
+        },
+        {
+          sprintName: sprint3Name,
+          value: {
+            blockedPercentage: 0.5,
+            developingPercentage: 0.5,
+          },
+        },
+      ],
+      {
+        totalBlockedPercentage: 0.5,
+        blockReasonPercentage: [
+          {
+            reasonName: JiraBlockReasonEnum.DEPENDENCIES_NOT_WORK,
+            percentage: 0.07,
+          },
+          {
+            reasonName: JiraBlockReasonEnum.TAKE_LEAVE,
+            percentage: 0.07,
+          },
+          {
+            reasonName: JiraBlockReasonEnum.OTHERS,
+            percentage: 0.07,
+          },
+        ],
+      }
+    );
+
+    expect(sprintStartistics).deep.equal(expected);
+  });
+  it("should return the empty Jira sprint statistics when statistic data is empty", () => {
+    const sprintBlockReasonPercentageMap = new Map<string, number>();
+    const sprintBlockPercentageMap = new Map<string, any>();
+    const sprintStandardDeviationMap = new Map<string, any>();
+    const sprintCompletedCardsNumberMap = new Map<string, number>();
+
+    const sprintStatistics = jiraTest.generateJiraSprintStatistics(
+      sprintCompletedCardsNumberMap,
+      sprintBlockPercentageMap,
+      sprintStandardDeviationMap,
+      sprintBlockReasonPercentageMap
+    );
+    const expected = new SprintStatistics([], [], [], {
+      totalBlockedPercentage: 0,
+      blockReasonPercentage: [],
+    });
+    expect(sprintStatistics).deep.equal(expected);
+  });
+});
+
+describe("sort sprints by sprint start date", () => {
+  it("sort sprints by sprint start date", () => {
+    const unorderedMap = new Map<string, JiraCardResponse[]>([
+      [sprint1Name, cardList1],
+      [sprint2Name, cardList2],
+    ]);
+    const excepted = new Map<string, JiraCardResponse[]>([
+      [sprint2Name, cardList2],
+      [sprint1Name, cardList1],
+    ]);
+
+    const orderedMap = jiraTest.sortBySprintStartDate(unorderedMap, sprints);
+    expect(orderedMap).deep.equal(excepted);
+  });
+});
+describe("get active and closed Sprints", () => {
+  it("get active and closed Sprints", () => {
+    const excepted = [activeSprint, closedSprint];
+    const filteredSprints = jiraTest.getActiveAndClosedSprints(sprints);
+    expect(filteredSprints).deep.equal(excepted);
+  });
+});
+
+describe("get sprint statistics", () => {
+  it("get sprint statistics", async () => {
+    sinon.stub(Jira.prototype, <any>"getAllSprintsByBoardId").returns(sprints);
+    const cards = { storyPointSum: 0, cardsNumber: 3, matchedCards: cardList1 };
+    const expected = new SprintStatistics(
+      [
+        { sprintName: "test Sprint 2", value: 1 },
+        { sprintName: "test Sprint 1", value: 1 },
+      ],
+      [
+        {
+          sprintName: "test Sprint 2",
+          value: { standardDeviation: 0, average: 5 },
+        },
+        {
+          sprintName: "test Sprint 1",
+          value: { standardDeviation: 0, average: 2 },
+        },
+      ],
+      [
+        {
+          sprintName: "test Sprint 2",
+          value: { blockedPercentage: 0.2, developingPercentage: 0.8 },
+        },
+        {
+          sprintName: "test Sprint 1",
+          value: { blockedPercentage: 0.5, developingPercentage: 0.5 },
+        },
+      ],
+      {
+        totalBlockedPercentage: 0.5,
+        blockReasonPercentage: [
+          { reasonName: "dependencies_not_work", percentage: 0.5 },
+          { reasonName: "sit_env_down", percentage: 0 },
+          { reasonName: "priority_change", percentage: 0 },
+          { reasonName: "solution_review", percentage: 0 },
+          { reasonName: "pr_review", percentage: 0 },
+          { reasonName: "question_to_be_answered", percentage: 0 },
+          { reasonName: "take_leave", percentage: 0 },
+          { reasonName: "others", percentage: 0 },
+        ],
+      }
+    );
+    const actual: SprintStatistics = await jira.getSprintStatistics(
+      storyPointsAndCycleTimeRequest,
+      cards
+    );
+    expect(expected).deep.equals(actual);
   });
 });
